Move utils to a new folder 'utils' and add history.js
