+++ src/utils/history.js
@@ -1,0 +1,2 @@
+import createHistory from 'history/createBrowserHistory'
+export default createHistory()
